# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERB Platform API - connecting suppliers with retailers in Saudi Arabia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (business owner)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/mine:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: List caller's own orders (business owner)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/assigned:
    get:
      operationId: getAssignedOrders
      tags: [orders]
      summary: List orders assigned to the calling supplier
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get one order (owner, assigned supplier, or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update delivery status (assigned supplier/carrier, owner cancel, or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdateInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/history:
    get:
      operationId: getOrderHistory
      tags: [orders]
      summary: Status history timeline for an order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderStatusHistory"
  /orders/{id}/offers:
    get:
      operationId: getOrderOffers
      tags: [orders]
      summary: Offer chain for an order (admin sees all; candidate sees own)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderOffer"
...
  /orders/{id}/offers/accept:
    post:
      operationId: acceptOrderOffer
      tags: [orders]
      summary: Accept an offer (supplier chooses delivery_mode)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                delivery_mode:
                  type: string
                  enum: [supplier_delivery, logistics]
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/offers/reject:
    post:
      operationId: rejectOrderOffer
      tags: [orders]
      summary: Reject current offer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
  /admin/orders:
    get:
      operationId: adminGetOrders
      tags: [admin]
      summary: List all orders (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /admin/orders/{id}/offers:
    get:
      operationId: adminGetOrderOffers
      tags: [admin]
      summary: Full candidate offer table for an order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offers with score breakdowns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderOffer"
  /admin/orders/{id}:
    patch:
      operationId: adminActOnOrder
      tags: [admin]
      summary: Approve, reject, force-assign, or resolve needs_manual (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOrderActionInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Order:
      type: object
      properties:
        id:
          type: integer
        business_id:
          type: string
        business_company:
          type: string
        business_contact:
          type: string
        business_phone:
          type: ["string", "null"]
        product_category:
          type: string
        items:
          type: string
        delivery_region:
          type: string
        delivery_address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum:
            [
              pending,
              approved,
              offered,
              assigned,
              preparing,
              ready_for_pickup,
              in_transit,
              delivered,
              rejected,
              cancelled,
              needs_manual,
            ]
        assigned_supplier_id:
          type: ["string", "null"]
        assigned_supplier_company:
          type: ["string", "null"]
        assigned_supplier_region:
          type: ["string", "null"]
        delivery_mode:
          type: ["string", "null"]
        assigned_logistics_id:
          type: ["string", "null"]
        assigned_logistics_company:
          type: ["string", "null"]
        offered_to_id:
          type: ["string", "null"]
        offer_expires_at:
          type: ["string", "null"]
        automation_score:
          type: ["string", "null"]
        picked_up_at:
          type: ["string", "null"]
        eta:
          type: ["string", "null"]
        admin_notes:
          type: ["string", "null"]
        ai_recommendation:
          type: ["string", "null"]
        automation_reason:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - business_id
        - business_company
        - business_contact
        - product_category
        - items
        - delivery_region
        - status
        - created_at
        - updated_at
    OrderCreateInput:
      type: object
      description: >
        Client must not send business_id, business_company, business_contact,
        or business_phone — those are filled from the authenticated profile.
      properties:
        product_category:
          type: string
        items:
          type: string
          minLength: 10
        delivery_region:
          type: string
        delivery_address:
          type: string
        notes:
          type: string
      required:
        - product_category
        - items
        - delivery_region
    OrderStatusUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [preparing, ready_for_pickup, in_transit, delivered, cancelled]
      required:
        - status
    OrderStatusHistory:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        from_status:
          type: ["string", "null"]
        to_status:
          type: string
        actor_id:
          type: ["string", "null"]
        actor_role:
          type: ["string", "null"]
        actor_kind:
          type: string
          enum: [user, system]
        note:
          type: ["string", "null"]
        created_at:
          type: string
    OrderOffer:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        candidate_id:
          type: string
        candidate_role:
          type: string
          enum: [supplier, logistics]
        rank:
          type: integer
        score:
          type: ["string", "null"]
        score_breakdown:
          type: object
          additionalProperties:
            type: number
        offered_at:
          type: string
        expires_at:
          type: string
        responded_at:
          type: ["string", "null"]
        response:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        created_at:
          type: string
...
    AdminOrderActionInput:
      type: object
      properties:
        action:
          type: string
          enum:
            [
              approve,
              reject,
              assign,
              force_assign,
              resolve,
              set_delivery_mode,
              run_automation,
            ]
        supplier_id:
          type: string
        logistics_id:
          type: string
        delivery_mode:
          type: string
          enum: [supplier_delivery, logistics]
        admin_notes:
          type: string
        reason:
          type: string
        target_status:
          type: string
          enum: [approved, offered, assigned, rejected, cancelled]
      required:
        - action
...